Fix Float-to-String coercion on Scala.js and fix scalafmt
- Use BigDecimal round-trip for Float-to-String coercion to avoid
  IEEE 754 precision artifacts on Scala.js (3.14f was producing
  "3.140000104904175" instead of "3.14")
- Run scalafmt on all migration source and test files

+++ schema/shared/src/main/scala/zio/blocks/schema/migration/DynamicMigration.scala
@@ -861,7 +861,7 @@ object DynamicMigration {
     case PrimitiveValue.Short(v)      => v.toString
     case PrimitiveValue.Int(v)        => v.toString
     case PrimitiveValue.Long(v)       => v.toString
-    case PrimitiveValue.Float(v)      => v.toString
+    case PrimitiveValue.Float(v)      => BigDecimal(v.toDouble).floatValue.toString
     case PrimitiveValue.Double(v)     => v.toString
     case PrimitiveValue.Char(v)       => v.toString
     case PrimitiveValue.BigInt(v)     => v.toString
